% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BB1copula.R
\name{BB1Copula}
\alias{BB1Copula}
\alias{r270BB1Copula}
\alias{r90BB1Copula}
\alias{surBB1Copula}
\title{Constructor of the BB1 Family and Rotated Versions thereof}
\usage{
BB1Copula(param = c(1, 1))
}
\arguments{
\item{param}{The parameter \code{param} defines the copula through
\code{theta} and \code{delta}.}
}
\value{
One of the respective BB1 copula classes
(\code{\linkS4class{BB1Copula}}, \code{\linkS4class{surBB1Copula}},
\code{\linkS4class{r90BB1Copula}}, \code{\linkS4class{r270BB1Copula}}).
}
\description{
Constructs an object of the \code{\linkS4class{BB1Copula}} (survival
\code{sur}, 90 degree rotated \code{r90} and 270 degree rotated \code{r270})
family for given parameters.
}
\examples{
library(copula)

persp(BB1Copula(c(1,1.5)), dCopula, zlim = c(0,10))
persp(surBB1Copula(c(1,1.5)), dCopula, zlim = c(0,10))
persp(r90BB1Copula(c(-1,-1.5)), dCopula, zlim = c(0,10))
persp(r270BB1Copula(c(-1,-1.5)), dCopula, zlim = c(0,10))
}
\author{
Benedikt Graeler
}
\references{
Joe, H., (1997). Multivariate Models and Dependence Concepts.
Monogra. Stat. Appl. Probab. 73, London: Chapman and Hall.
}
\seealso{
See also \code{\link{BB6Copula}}, \code{\link{BB7Copula}},
\code{\link{BB8Copula}} and \code{\link{joeCopula}} for further wrapper
functions to the \code{\link{VineCopula-package}}.
}
\keyword{copula}
\keyword{distribution}

