% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC2copula-depr.R
\docType{class}
\name{VC2copula-deprecated}
\alias{VC2copula-deprecated}
\alias{copulaFromFamilyIndex}
\alias{surClaytonCopula-class}
\alias{surClaytonCopula}
\alias{dduCopula,matrix,surClaytonCopula-method}
\alias{dduCopula,numeric,surClaytonCopula-method}
\alias{ddvCopula,matrix,surClaytonCopula-method}
\alias{ddvCopula,numeric,surClaytonCopula-method}
\alias{r90ClaytonCopula-class}
\alias{dduCopula,matrix,r90ClaytonCopula-method}
\alias{dduCopula,numeric,r90ClaytonCopula-method}
\alias{ddvCopula,matrix,r90ClaytonCopula-method}
\alias{ddvCopula,numeric,r90ClaytonCopula-method}
\alias{r270ClaytonCopula-class}
\alias{dduCopula,matrix,r270ClaytonCopula-method}
\alias{dduCopula,numeric,r270ClaytonCopula-method}
\alias{ddvCopula,matrix,r270ClaytonCopula-method}
\alias{ddvCopula,numeric,r270ClaytonCopula-method}
\alias{surGumbelCopula-class}
\alias{surGumbelCopula}
\alias{dduCopula,matrix,surGumbelCopula-method}
\alias{dduCopula,numeric,surGumbelCopula-method}
\alias{ddvCopula,matrix,surGumbelCopula-method}
\alias{ddvCopula,numeric,surGumbelCopula-method}
\alias{r90GumbelCopula-class}
\alias{dduCopula,matrix,r90GumbelCopula-method}
\alias{dduCopula,numeric,r90GumbelCopula-method}
\alias{ddvCopula,matrix,r90GumbelCopula-method}
\alias{ddvCopula,numeric,r90GumbelCopula-method}
\alias{r270GumbelCopula-class}
\alias{dduCopula,matrix,r270GumbelCopula-method}
\alias{dduCopula,numeric,r270GumbelCopula-method}
\alias{ddvCopula,matrix,r270GumbelCopula-method}
\alias{ddvCopula,numeric,r270GumbelCopula-method}
\alias{joeBiCopula-class}
\alias{joeBiCopula}
\alias{dduCopula,numeric,joeBiCopula-method}
\alias{ddvCopula,numeric,joeBiCopula-method}
\alias{dduCopula,matrix,joeBiCopula-method}
\alias{ddvCopula,matrix,joeBiCopula-method}
\alias{getKendallDistr,joeBiCopula-method}
\alias{kendallDistribution,joeBiCopula-method}
\alias{surJoeBiCopula-class}
\alias{dduCopula,numeric,surJoeBiCopula-method}
\alias{ddvCopula,numeric,surJoeBiCopula-method}
\alias{dduCopula,matrix,surJoeBiCopula-method}
\alias{ddvCopula,matrix,surJoeBiCopula-method}
\alias{r90JoeBiCopula-class}
\alias{dduCopula,numeric,r90JoeBiCopula-method}
\alias{ddvCopula,numeric,r90JoeBiCopula-method}
\alias{dduCopula,matrix,r90JoeBiCopula-method}
\alias{ddvCopula,matrix,r90JoeBiCopula-method}
\alias{r270JoeBiCopula-class}
\alias{dduCopula,numeric,r270JoeBiCopula-method}
\alias{ddvCopula,numeric,r270JoeBiCopula-method}
\alias{dduCopula,matrix,r270JoeBiCopula-method}
\alias{ddvCopula,matrix,r270JoeBiCopula-method}
\alias{BB1Copula-class}
\alias{BB1Copula}
\alias{dduCopula,numeric,BB1Copula-method}
\alias{ddvCopula,numeric,BB1Copula-method}
\alias{dduCopula,matrix,BB1Copula-method}
\alias{ddvCopula,matrix,BB1Copula-method}
\alias{getKendallDistr,BB1Copula-method}
\alias{kendallDistribution,BB1Copula-method}
\alias{surBB1Copula-class}
\alias{dduCopula,numeric,surBB1Copula-method}
\alias{ddvCopula,numeric,surBB1Copula-method}
\alias{dduCopula,matrix,surBB1Copula-method}
\alias{ddvCopula,matrix,surBB1Copula-method}
\alias{r90BB1Copula-class}
\alias{dduCopula,numeric,r90BB1Copula-method}
\alias{ddvCopula,numeric,r90BB1Copula-method}
\alias{dduCopula,matrix,r90BB1Copula-method}
\alias{ddvCopula,matrix,r90BB1Copula-method}
\alias{r270BB1Copula-class}
\alias{dduCopula,numeric,r270BB1Copula-method}
\alias{ddvCopula,numeric,r270BB1Copula-method}
\alias{dduCopula,matrix,r270BB1Copula-method}
\alias{ddvCopula,matrix,r270BB1Copula-method}
\alias{BB6Copula-class}
\alias{BB6Copula}
\alias{dduCopula,numeric,BB6Copula-method}
\alias{ddvCopula,numeric,BB6Copula-method}
\alias{dduCopula,matrix,BB6Copula-method}
\alias{ddvCopula,matrix,BB6Copula-method}
\alias{getKendallDistr,BB6Copula-method}
\alias{kendallDistribution,BB6Copula-method}
\alias{surBB6Copula-class}
\alias{dduCopula,numeric,surBB6Copula-method}
\alias{ddvCopula,numeric,surBB6Copula-method}
\alias{dduCopula,matrix,surBB6Copula-method}
\alias{ddvCopula,matrix,surBB6Copula-method}
\alias{r90BB6Copula-class}
\alias{dduCopula,numeric,r90BB6Copula-method}
\alias{ddvCopula,numeric,r90BB6Copula-method}
\alias{dduCopula,matrix,r90BB6Copula-method}
\alias{ddvCopula,matrix,r90BB6Copula-method}
\alias{r270BB6Copula-class}
\alias{dduCopula,numeric,r270BB6Copula-method}
\alias{ddvCopula,numeric,r270BB6Copula-method}
\alias{dduCopula,matrix,r270BB6Copula-method}
\alias{ddvCopula,matrix,r270BB6Copula-method}
\alias{BB7Copula-class}
\alias{BB7Copula}
\alias{dduCopula,numeric,BB7Copula-method}
\alias{ddvCopula,numeric,BB7Copula-method}
\alias{dduCopula,matrix,BB7Copula-method}
\alias{ddvCopula,matrix,BB7Copula-method}
\alias{getKendallDistr,BB7Copula-method}
\alias{kendallDistribution,BB7Copula-method}
\alias{surBB7Copula-class}
\alias{dduCopula,numeric,surBB7Copula-method}
\alias{ddvCopula,numeric,surBB7Copula-method}
\alias{dduCopula,matrix,surBB7Copula-method}
\alias{ddvCopula,matrix,surBB7Copula-method}
\alias{r90BB7Copula-class}
\alias{dduCopula,numeric,r90BB7Copula-method}
\alias{ddvCopula,numeric,r90BB7Copula-method}
\alias{dduCopula,matrix,r90BB7Copula-method}
\alias{ddvCopula,matrix,r90BB7Copula-method}
\alias{r270BB7Copula-class}
\alias{dduCopula,numeric,r270BB7Copula-method}
\alias{ddvCopula,numeric,r270BB7Copula-method}
\alias{dduCopula,matrix,r270BB7Copula-method}
\alias{ddvCopula,matrix,r270BB7Copula-method}
\alias{BB8Copula-class}
\alias{BBB8Copula}
\alias{B8Copula-class}
\alias{dduCopula,numeric,BB8Copula-method}
\alias{ddvCopula,numeric,BB8Copula-method}
\alias{dduCopula,matrix,BB8Copula-method}
\alias{ddvCopula,matrix,BB8Copula-method}
\alias{getKendallDistr,BB8Copula-method}
\alias{kendallDistribution,BB8Copula-method}
\alias{surBB8Copula-class}
\alias{dduCopula,numeric,surBB8Copula-method}
\alias{ddvCopula,numeric,surBB8Copula-method}
\alias{dduCopula,matrix,surBB8Copula-method}
\alias{ddvCopula,matrix,surBB8Copula-method}
\alias{r90BB8Copula-class}
\alias{dduCopula,numeric,r90BB8Copula-method}
\alias{ddvCopula,numeric,r90BB8Copula-method}
\alias{dduCopula,matrix,r90BB8Copula-method}
\alias{ddvCopula,matrix,r90BB8Copula-method}
\alias{r270BB8Copula-class}
\alias{dduCopula,numeric,r270BB8Copula-method}
\alias{ddvCopula,numeric,r270BB8Copula-method}
\alias{dduCopula,matrix,r270BB8Copula-method}
\alias{ddvCopula,matrix,r270BB8Copula-method}
\alias{tawnT1Copula-class}
\alias{tawnT1Copula}
\alias{dduCopula,matrix,tawnT1Copula-method}
\alias{dduCopula,numeric,tawnT1Copula-method}
\alias{ddvCopula,matrix,tawnT1Copula-method}
\alias{ddvCopula,numeric,tawnT1Copula-method}
\alias{surTawnT1Copula-class}
\alias{dduCopula,matrix,surTawnT1Copula-method}
\alias{dduCopula,numeric,surTawnT1Copula-method}
\alias{ddvCopula,matrix,surTawnT1Copula-method}
\alias{ddvCopula,numeric,surTawnT1Copula-method}
\alias{r90TawnT1Copula-class}
\alias{dduCopula,matrix,r90TawnT1Copula-method}
\alias{dduCopula,numeric,r90TawnT1Copula-method}
\alias{ddvCopula,matrix,r90TawnT1Copula-method}
\alias{ddvCopula,numeric,r90TawnT1Copula-method}
\alias{r270TawnT1Copula-class}
\alias{dduCopula,matrix,r270TawnT1Copula-method}
\alias{dduCopula,numeric,r270TawnT1Copula-method}
\alias{ddvCopula,matrix,r270TawnT1Copula-method}
\alias{ddvCopula,numeric,r270TawnT1Copula-method}
\alias{tawnT2Copula-class}
\alias{tawnT2Copula}
\alias{dduCopula,matrix,tawnT2Copula-method}
\alias{dduCopula,numeric,tawnT2Copula-method}
\alias{ddvCopula,matrix,tawnT2Copula-method}
\alias{ddvCopula,numeric,tawnT2Copula-method}
\alias{surTawnT2Copula-class}
\alias{dduCopula,matrix,surTawnT2Copula-method}
\alias{dduCopula,numeric,surTawnT2Copula-method}
\alias{ddvCopula,matrix,surTawnT2Copula-method}
\alias{ddvCopula,numeric,surTawnT2Copula-method}
\alias{r90TawnT2Copula-class}
\alias{dduCopula,matrix,r90TawnT2Copula-method}
\alias{dduCopula,numeric,r90TawnT2Copula-method}
\alias{ddvCopula,matrix,r90TawnT2Copula-method}
\alias{ddvCopula,numeric,r90TawnT2Copula-method}
\alias{r270TawnT2Copula-class}
\alias{dduCopula,matrix,r270TawnT2Copula-method}
\alias{dduCopula,numeric,r270TawnT2Copula-method}
\alias{ddvCopula,matrix,r270TawnT2Copula-method}
\alias{ddvCopula,numeric,r270TawnT2Copula-method}
\alias{vineCopula-class}
\alias{vineCopula}
\alias{fitCopula}
\alias{vineCopula-method}
\title{Deprecated}
\usage{
copulaFromFamilyIndex(family, par, par2 = 0)
}
\arguments{
\item{family}{..}

\item{par}{...}

\item{par2}{...}

\item{RVM}{...}

\item{type}{...}

\item{param}{...}
}
\description{
This functionality is deprecated in 'VineCopula'. Use the package 'VC2copula'
instead.
}
