% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vineCopulas.R
\docType{class}
\name{vineCopula-class}
\alias{fitCopula,vineCopula-method}
\alias{vineCopula-class}
\title{Class \code{"vineCopula"}}
\description{
A class representing vine copulas in a object oriented implementations. Many
functions go back to the package \code{\link{VineCopula-package}}
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("vineCopula", ...)} or through the function
\code{\link{vineCopula}}.
}
\examples{

showClass("vineCopula")

}
\author{
Benedikt Graeler
}
\references{
Aas, K., C. Czado, A. Frigessi, and H. Bakken (2009).
Pair-copula constructions of multiple dependence Insurance: Mathematics and
Economics 44 (2), 182-198.
}
\seealso{
\code{\link{RVineMatrix}} from package
\code{\link{VineCopula-package}}
}
\keyword{classes}

