% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB6copula.R
\docType{class}
\name{BB6Copula-class}
\alias{BB6Copula-class}
\alias{dduCopula,matrix,BB6Copula-method}
\alias{dduCopula,matrix,r270BB6Copula-method}
\alias{dduCopula,matrix,r90BB6Copula-method}
\alias{dduCopula,matrix,surBB6Copula-method}
\alias{dduCopula,numeric,BB6Copula-method}
\alias{dduCopula,numeric,r270BB6Copula-method}
\alias{dduCopula,numeric,r90BB6Copula-method}
\alias{dduCopula,numeric,surBB6Copula-method}
\alias{ddvCopula,matrix,BB6Copula-method}
\alias{ddvCopula,matrix,r270BB6Copula-method}
\alias{ddvCopula,matrix,r90BB6Copula-method}
\alias{ddvCopula,matrix,surBB6Copula-method}
\alias{ddvCopula,numeric,BB6Copula-method}
\alias{ddvCopula,numeric,r270BB6Copula-method}
\alias{ddvCopula,numeric,r90BB6Copula-method}
\alias{ddvCopula,numeric,surBB6Copula-method}
\alias{getKendallDistr,BB6Copula-method}
\alias{kendallDistribution,BB6Copula-method}
\alias{r270BB6Copula-class}
\alias{r90BB6Copula-class}
\alias{surBB6Copula-class}
\title{Classes \code{"BB6Copula"}, \code{"surBB6Copula"}, \code{"r90BB6Copula"} and
\code{"r270BB6Copula"}}
\description{
Wrapper classes representing the BB6, survival BB6, 90 degree and 270 degree
rotated BB6 copula families (Joe 1997) from the
\code{\link{VineCopula-package}}.
}
\section{Objects from the Classes}{
 Objects can be created by calls of the
form \code{new("BB6Copula", ...)}, \code{new("surBB6Copula", ...)},
\code{new("r90BB6Copula", ...)} and \code{new("r270BB6Copula", ...)} or by
the functions \code{\link{BB6Copula}}, \code{\link{surBB6Copula}},
\code{\link{r90BB6Copula}} and \code{\link{r270BB6Copula}}.
}
\examples{

showClass("BB6Copula")

}
\author{
Benedikt Graeler
}
\references{
Joe, H., (1997). Multivariate Models and Dependence Concepts.
Monogra. Stat. Appl. Probab. 73, London: Chapman and Hall.
}
\seealso{
See also \code{\linkS4class{BB1Copula}},
\code{\linkS4class{BB7Copula}}, \code{\linkS4class{BB8Copula}} and
\code{\linkS4class{joeCopula}} for further wrapper classes to the
\code{\link{VineCopula-package}}.
}
\keyword{classes}

