% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB8copula.R
\docType{class}
\name{BB8Copula-class}
\alias{BB8Copula-class}
\alias{dduCopula,numeric,BB8Copula-method}
\alias{ddvCopula,numeric,BB8Copula-method}
\alias{dduCopula,matrix,BB8Copula-method}
\alias{ddvCopula,matrix,BB8Copula-method}
\alias{getKendallDistr,BB8Copula-method}
\alias{kendallDistribution,BB8Copula-method}
\alias{surBB8Copula-class}
\alias{dduCopula,numeric,surBB8Copula-method}
\alias{ddvCopula,numeric,surBB8Copula-method}
\alias{dduCopula,matrix,surBB8Copula-method}
\alias{ddvCopula,matrix,surBB8Copula-method}
\alias{r90BB8Copula-class}
\alias{dduCopula,numeric,r90BB8Copula-method}
\alias{ddvCopula,numeric,r90BB8Copula-method}
\alias{dduCopula,matrix,r90BB8Copula-method}
\alias{ddvCopula,matrix,r90BB8Copula-method}
\alias{r270BB8Copula-class}
\alias{dduCopula,numeric,r270BB8Copula-method}
\alias{ddvCopula,numeric,r270BB8Copula-method}
\alias{dduCopula,matrix,r270BB8Copula-method}
\alias{ddvCopula,matrix,r270BB8Copula-method}
\alias{fitCopula,twoParamBiCop-method}
\title{Classes \code{"BB8Copula"}, \code{"surBB8Copula"}, \code{"r90BB8Copula"} and
\code{"r270BB8Copula"}}
\description{
Wrapper classes representing the BB8, survival BB8, 90 degree and 270 degree
rotated BB8 copula families (Joe 1997) from the
\code{\link{VineCopula-package}} package.
}
\section{Objects from the Classes}{
 Objects can be created by calls of the
form \code{new("BB8Copula", ...)}, \code{new("surBB8Copula", ...)},
\code{new("r90BB8Copula", ...)} and \code{new("r270BB8Copula", ...)} or by
the functions \code{\link{BB8Copula}}, \code{\link{surBB8Copula}},
\code{\link{r90BB8Copula}} and \code{\link{r270BB8Copula}}.
}

\examples{

showClass("BB8Copula")

}
\references{
Joe, H., (1997). Multivariate Models and Dependence Concepts.
Monogra. Stat. Appl. Probab. 73, London: Chapman and Hall.
}
\seealso{
See also \code{\linkS4class{BB1Copula}},
\code{\linkS4class{BB6Copula}}, \code{\linkS4class{BB7Copula}} and
\code{\linkS4class{joeCopula}} for further wrapper classes to the
\code{\link{VineCopula-package}}.
}
\author{
Benedikt Graeler
}
\keyword{classes}
