% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VineCopula-package.R
\docType{data}
\name{daxreturns}
\alias{daxreturns}
\title{Major German Stocks}
\format{A data frame with 1158 observations on 15 variables. Column names
correspond to ticker symbols of the stocks.}
\source{
Yahoo! Finance
}
\description{
This data set contains transformed standardized residuals of daily log
returns of 15 major German stocks represented in the index DAX observed from
January 2005 to August 2009. Each time series is filtered using a GARCH(1,1)
model with Student t innovations.
}
\examples{

# load the data set
data(daxreturns)

# compute the empirical Kendall's tau matrix
TauMatrix(daxreturns)

}
\seealso{
\code{\link[=RVineStructureSelect]{RVineStructureSelect()}}
}
