% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeBiCopula.R
\docType{class}
\name{joeBiCopula-class}
\alias{joeBiCopula-class}
\alias{dduCopula,numeric,joeBiCopula-method}
\alias{ddvCopula,numeric,joeBiCopula-method}
\alias{dduCopula,matrix,joeBiCopula-method}
\alias{ddvCopula,matrix,joeBiCopula-method}
\alias{getKendallDistr,joeBiCopula-method}
\alias{kendallDistribution,joeBiCopula-method}
\alias{surJoeBiCopula-class}
\alias{dduCopula,numeric,surJoeBiCopula-method}
\alias{ddvCopula,numeric,surJoeBiCopula-method}
\alias{dduCopula,matrix,surJoeBiCopula-method}
\alias{ddvCopula,matrix,surJoeBiCopula-method}
\alias{r90JoeBiCopula-class}
\alias{dduCopula,numeric,r90JoeBiCopula-method}
\alias{ddvCopula,numeric,r90JoeBiCopula-method}
\alias{dduCopula,matrix,r90JoeBiCopula-method}
\alias{ddvCopula,matrix,r90JoeBiCopula-method}
\alias{r270JoeBiCopula-class}
\alias{dduCopula,numeric,r270JoeBiCopula-method}
\alias{ddvCopula,numeric,r270JoeBiCopula-method}
\alias{dduCopula,matrix,r270JoeBiCopula-method}
\alias{ddvCopula,matrix,r270JoeBiCopula-method}
\title{Classes \code{"joeBiCopula"}, \code{"surJoeBiCopula"},
\code{"r90JoeBiCopula"} and \code{"r270JoeBiCopula"}}
\description{
Wrapper classes representing the bivariate Joe, survival Joe, 90 degree and
270 degree rotated Joe copula families (Joe 1997) from
\code{\link{VineCopula-package}}. Note that package
\code{\link{copula-package}} provides a class \code{\linkS4class{joeCopula}}
as well.
}
\section{Objects from the Classes}{
 Objects can be created by calls of the
form \code{new("joeBiCopula", ...)}, \code{new("surJoeBiCopula", ...)},
\code{new("r90JoeBiCopula", ...)} and \code{new("r270JoeBiCopula", ...)} or
by the functions \code{\link{joeBiCopula}}, \code{\link{surJoeBiCopula}},
\code{\link{r90JoeBiCopula}} and \code{\link{r270JoeBiCopula}}.
}

\examples{

showClass("surJoeBiCopula")

}
\references{
Joe, H., (1997). Multivariate Models and Dependence Concepts.
Monogra. Stat. Appl. Probab. 73, London: Chapman and Hall.
}
\seealso{
See also \code{\linkS4class{BB1Copula}},
\code{\linkS4class{BB6Copula}}, \code{\linkS4class{BB7Copula}} and
\code{\linkS4class{BB8Copula}} for further wrapper classes to the
\code{\link{VineCopula-package}}.
}
\author{
Benedikt Graeler
}
\keyword{classes}
