% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVineMatrixSample.R
\name{RVineMatrixSample}
\alias{RVineMatrixSample}
\title{R-Vine Matrix Sample}
\usage{
RVineMatrixSample(d, size = 5, naturalOrder = FALSE)
}
\arguments{
\item{d}{Dimension of the R-Vine matrices.}

\item{size}{Number of matrices to sample.}

\item{naturalOrder}{Should the matrices be in the natural order
(default: \code{naturalOrder = FALSE}).}
}
\value{
A list of length \code{size} with each element containing one
R-Vine matrix.
}
\description{
Sample R-Vine matrices.
}
\examples{

# Matrix and sample sizes
d <- 10
size <- 5

# Sample in the natural order
RVM <- RVineMatrixSample(d, size)
sapply(RVM, RVineMatrixCheck)

# Sample without natural order
RVM <- RVineMatrixSample(d, size, naturalOrder = TRUE)
sapply(RVM, RVineMatrixCheck)

}
\author{
Thibault Vatter
}
\references{
Joe H, Cooke RM and Kurowicka D (2011). Regular vines:
generation algorithm and number of equivalence classes. In Dependence
Modeling: Vine Copula Handbook, pp 219--231. World Scientific, Singapore.
}
\seealso{
\code{\link{RVineMatrix}}, \code{\link{RVineMatrixCheck}}
}

