% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tawnCopula.R
\docType{class}
\name{tawnT1Copula-class}
\alias{tawnT1Copula-class}
\alias{dduCopula,matrix,tawnT1Copula-method}
\alias{dduCopula,numeric,tawnT1Copula-method}
\alias{ddvCopula,matrix,tawnT1Copula-method}
\alias{ddvCopula,numeric,tawnT1Copula-method}
\alias{surTawnT1Copula-class}
\alias{dduCopula,matrix,surTawnT1Copula-method}
\alias{dduCopula,numeric,surTawnT1Copula-method}
\alias{ddvCopula,matrix,surTawnT1Copula-method}
\alias{ddvCopula,numeric,surTawnT1Copula-method}
\alias{r90TawnT1Copula-class}
\alias{dduCopula,matrix,r90TawnT1Copula-method}
\alias{dduCopula,numeric,r90TawnT1Copula-method}
\alias{ddvCopula,matrix,r90TawnT1Copula-method}
\alias{ddvCopula,numeric,r90TawnT1Copula-method}
\alias{r270TawnT1Copula-class}
\alias{dduCopula,matrix,r270TawnT1Copula-method}
\alias{dduCopula,numeric,r270TawnT1Copula-method}
\alias{ddvCopula,matrix,r270TawnT1Copula-method}
\alias{ddvCopula,numeric,r270TawnT1Copula-method}
\title{Class \code{"tawnT1Copula"}}
\description{
S4-class representation of the Tawn Copula family of type 1 and rotated
versions there of.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("tawnT1Copula", ...)}, or through the explicit constructors
\code{\link{tawnT1Copula}}, \code{\link{surTawnT1Copula}},
\code{\link{r90TawnT1Copula}} and \code{\link{r270TawnT1Copula}}
respectively.
}

\examples{

showClass("tawnT1Copula")

}
\seealso{
\code{\linkS4class{tawnT2Copula}} and the package
\code{\link{VineCopula-package}} for implementation details.
}
\author{
Benedikt Graeler
}
\keyword{classes}
